Minor (#86)
* Small fixes

* Allow zero size match

This was previously not allowed as it was considered bad to match
nothing.  However, if you change mode because you matched nothing then
that is not possible without this change.  For instance in a white space
sensitive language you might do
```
  p("start",
    {
        "[[:blank:]]*" >>
            [set_indent](auto &m) {
              m.mode("main");
              set_indent(m, m.match(1).len);
            },
    });

  p("main",
    {
        ...
        // On new line go back to finding indent
        "\\n" >> [](auto &m) { m.mode("start"); },
    }
```

* Make get_errors less clever.

The get_errors function was using a clever move semantics, but it was not
working on Clang. So removing.

* Code review